feat(@angular-devkit/build-angular): differential loading
This PR adds differential loading to the browser builder. First, it checks if differential loading is needed. This is the case if the compilation target is ES2015 while the browserslist points to ES5 browsers.

For providing differential loading, it calls the methods for creating the webpack config for each compilation target (e. g. ES5 and ES2015). The needed differences between those configurations are defined using parameters also added by this PR. Then it calls webpack for each of them, merges the results and writes the index.html.

This feature is currently hidden behind a flag. To activate it, set the ENABLE_DIFFERENTIAL_LOADING flag in packages/angular_devkit/core/src/index.ts