feat(@schematics/angular): introduce Ivy workspace concept
At the moment, to enable Ivy as the view rendering engine we are adding `enableIvy: true` in each of the applications tsconfigs.
While this works this is not ideal when users want to opt-out, as users will need to modify multiple tsconfig.

With the introduction of Ivy libraries for development this will be even harder and cumbersome.

We added the `enableIvy` option to the `workspace` schematic and remove the option for the `application` schematic.

`ng new <project-name> --enable-ivy` will pass down down the option the workspace schematic and an Ivy workspace will be created.
Other schematics such as `application` and `library` will read the workspace (root) tsconfig to verify if a application or library with
 Ivy structure and configuration should be created.